+++ backend/googleAuth/Dockerfile
@@ -7,6 +7,8 @@ WORKDIR /app/
 RUN mkdir api public
 
 COPY package*.json /app/
+COPY ssl/cert.pem /app/ssl/cert.pem
+COPY ssl/key.pem /app/ssl/key.pem
 RUN npm install --production
 
 COPY api /app/api/
