--- frontend/env.development_example
@@ -1,2 +1,0 @@
-VITE_API_USER="http://localhost:3002/api"
-VITE_API_GAME="http://localhost:3001/api"
