--- frontend/env.production_example
@@ -1,2 +1,0 @@
-VITE_API_GAME="https://localhost:4433/games"
-VITE_API_USER="https://localhost:4433/user"
